% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Supply_Transition.R
\name{supply_transition}
\alias{supply_transition}
\title{Supply Transition function to convert the LOCAL CODE counts into ONET counts}
\usage{
supply_transition(
  country_supply_file = NA,
  state_supply_file = NA,
  msa_supply_file = NA,
  redist_SOC = 6,
  output_path = "",
  country_id = NA,
  sum_up_msa = "no",
  threshold_msa = "yes",
  username = NA,
  password = NA,
  host = NA
)
}
\arguments{
\item{county_supply_file:}{Counts for local code at country level. There should be a column named occupation_code, country_id, year and tot_emp.}

\item{state_supply_file:}{Counts for local code at state level. There should be a column named occupation_code, state_id, year and tot_emp.}

\item{msa_supply_file:}{Counts for local code at msa level. There should be a column named occupation_code, msa_id, year and tot_emp.}

\item{redist_SOC:}{The variable to facilitate redistributing uing demand data in case of same values at higher hierarchy. DEFAULT to 6}

\item{output_path:}{The path where all the output files will be saved.}

\item{country_id:}{The country_id of the country to get the geographical mapping from the database.}

\item{sum_up_msa:}{If MSAs cover the whole territory of a country then "yes". DEFAULT to "no"}

\item{threshold_msa:}{If the sum of the msa should not exceed state values then "yes". DEFAULT to "yes"}

\item{username:}{Gartner username}

\item{password:}{Workbench database password}

\item{host:}{The connection string to connect to the host}
}
\value{
All the files yearwise and otherwise will be wirtten on the output folder specified. A list will be returned with all the data too.\cr
final_list[[1]][[1]]: Country final files
final_list[[1]][[2]]: Country Rate
final_list[[2]][[1]]: State final files
final_list[[2]][[2]]: State Rate
final_list[[3]][[1]]: MSA final files
final_list[[3]][[2]]: MSA Rate
}
\description{
This function helps to convert the counts for LOCAL CODE into the ONET counts.
If the count at soc 6 level is same, 
then it is redistributed on the basis of the supply or demand file for the Benchmark country.
Here as a default value to redistribute SOC 6 is given, This can be changed in the function call.\cr\cr
}
\details{
All the required files will be asked to be inputed in the function.
Please read all the window titles to input correct files.

MSA adjustment can be optionally done.
}
